+++ test/example_struct/bmp2bin_rgb_struct.c
@@ -108,8 +108,8 @@ int main(int argc, const char *argv[])
     fwrite(&img->height, sizeof(unsigned int), 1, result);
 
 
-    unsigned int pixels_len = (img->width * img->height) * sizeof(rgb);
-    fwrite(&pixels_len, sizeof(unsigned int), 1, result);
+    unsigned long pixels_len = (img->width * img->height) * sizeof(rgb);
+    fwrite(&pixels_len, sizeof(unsigned long), 1, result);
 
 
     fwrite(img->pixels, pixels_len, 1, result);
